# perf: Reduce the use of temporary objects