--- a/packages/babel-generator/src/generators/jsx.ts
+++ b/packages/babel-generator/src/generators/jsx.ts
@@ -85,7 +85,7 @@ export function JSXOpeningElement(this: Printer, node: t.JSXOpeningElement) {
   this.print(node.typeParameters); // TS
   if (node.attributes.length > 0) {
     this.space();
-    this.printJoin(node.attributes, { separator: spaceSeparator });
+    this.printJoin(node.attributes, undefined, undefined, spaceSeparator);
   }
   if (node.selfClosing) {
     this.space();